fix: restructure project folder
