+++ swagger_output.json
@@ -2,18 +2,34 @@
   "swagger": "2.0",
   "info": {
     "version": "1.0.0",
-    "title": "REST API",
-    "description": ""
+    "title": "XIMI Server API",
+    "description": "API documentation list for project XIMI"
   },
   "host": "localhost:3000",
   "basePath": "/",
+  "tags": [
+    {
+      "name": "Rooms",
+      "description": "List of endpoints related to room"
+    }
+  ],
   "schemes": [
-    "http"
+    "http",
+    "https"
+  ],
+  "consumes": [
+    "application/json"
+  ],
+  "produces": [
+    "application/json"
   ],
   "paths": {
     "/rooms/create": {
-      "get": {
-        "description": "",
+      "post": {
+        "tags": [
+          "Rooms"
+        ],
+        "description": "Send a request to create a new room",
         "parameters": [],
         "responses": {
           "200": {
@@ -24,7 +40,10 @@
     },
     "/rooms/list": {
       "get": {
-        "description": "",
+        "tags": [
+          "Rooms"
+        ],
+        "description": "Send a request to fetch list of available rooms",
         "parameters": [],
         "responses": {
           "200": {
